[CMake] Rewrite FindLibSoup2.cmake. https://bugs.webkit.org/show_bug.cgi?id=93191
Reviewed by Rob Buis.

.:

The existing LibSoup2 was imported from somewhere else and not
only did it contain a lot of unnecessary cruft to look for libsoup
2.2 (which we do not support anyway), but it also relied on the
paths returned by pkg-config for setting the library and include
paths.

For one, this meant "-lsoup-2.4" was passed to the linked instead
of "-L/path/to/libsoup-2.4.so", which would sometimes make a
system version of libsoup to be picked up instead of the one
installed by, say, jhbuild.

The new FindLibSoup.cmake now only looks for libsoup 2.4 and
relies on pkg-config solely for retrieving the current LibSoup
version.

* Source/cmake/FindLibSoup.cmake: Added.
* Source/cmake/FindLibSoup2.cmake: Removed.
* Source/cmake/OptionsEfl.cmake: Look for LibSoup instead of
LibSoup2, look for the GObject component of Glib.

Source/WebCore:

* PlatformEfl.cmake: Use LIBSOUP_FOO instead of LIBSOUP24_FOO,
also link against GObject, which is needed by libsoup.

Source/WebKit:

* PlatformEfl.cmake: Use LIBSOUP_FOO instead of LIBSOUP24_FOO.

Source/WebKit2:

* PlatformEfl.cmake: Use LIBSOUP_FOO instead of LIBSOUP24_FOO, and
do not use LIBSOUP24_LDFLAGS, as it is not needed anymore.

Tools:

* DumpRenderTree/efl/CMakeLists.txt: Use LIBSOUP_FOO instead of
LIBSOUP24_FOO, and do not use LIBSOUP24_LDFLAGS, as it is not
needed anymore.
* EWebLauncher/CMakeLists.txt: Ditto.
* MiniBrowser/efl/CMakeLists.txt: Ditto.
* TestWebKitAPI/PlatformEfl.cmake: Ditto.
* WebKitTestRunner/CMakeLists.txt: Ditto.
* WebKitTestRunner/PlatformEfl.cmake: Ditto.

git-svn-id: svn://svn.chromium.org/blink/trunk@125467 bbb929c8-8fbe-4397-9dbb-9b2b20218538