Remove RefPtr from HTMLProgressElement::m_value https://bugs.webkit.org/show_bug.cgi?id=94336
Reviewed by Kent Tamura.

To avoid reference cycles of RefPtr<Node>s, we want to remove
unnecessary RefPtr<Node>s. The rationale is described in bug 94324.

HTMLProgressElement::m_value does not need to be a RefPtr<Node>, because
it is guaranteed to point to a shadow DOM tree of the HTMLProgressElement
node, which is guaranteed to exist in the subtree of the HTMLProgressElement node.

No tests. No change in behavior.

* html/HTMLProgressElement.cpp:
(WebCore::HTMLProgressElement::HTMLProgressElement):
(WebCore::HTMLProgressElement::createShadowSubtree):
* html/HTMLProgressElement.h:


git-svn-id: svn://svn.chromium.org/blink/trunk@125985 bbb929c8-8fbe-4397-9dbb-9b2b20218538