Crash in RenderTableSection::setCellLogicalWidths https://bugs.webkit.org/show_bug.cgi?id=94291
Reviewed by Abhishek Arya.

Source/WebCore:

This issue was that splitAnonymousBoxesAroundChild would move a table section
into a newly created table *without* marking it as needing cell recalc. The table
would thus never build its structure to match its sections. The fix is to hop on
the new willBeRemovedFromTree signal so that the section invalidates itself properly.

Test: fast/table/crash-split-table-section-no-cell-recalc.html

* rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::willBeRemovedFromTree):
* rendering/RenderTableSection.h:
Replaced willBeDestroyed by willBeRemovedFromTree in RenderTableSection. This ensures that it is called
when moving sections in the tree to mark them as needing cell recalc.

LayoutTests:

* fast/table/crash-split-table-section-no-cell-recalc-expected.txt: Added.
* fast/table/crash-split-table-section-no-cell-recalc.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@126251 bbb929c8-8fbe-4397-9dbb-9b2b20218538