[webkit-patch] gets stuck into an infinite loop if bugzilla doesn't respond in time. https://bugs.webkit.org/show_bug.cgi?id=94700
Patch by Szilard Ledan <szledan@inf.u-szeged.hu> on 2012-08-28
Reviewed by Eric Seidel.

Interrupting download with a timeout would be a possible solution.
I imported the socket module and used the setdefaulttimeout() method.
Fixing it is important, because EWS bots need to be restarted regularly
because of this bug.

* Scripts/webkitpy/common/net/bugzilla/bugzilla.py:
(Bugzilla._get_browser):
(Bugzilla.setdefaulttimeout):

git-svn-id: svn://svn.chromium.org/blink/trunk@126905 bbb929c8-8fbe-4397-9dbb-9b2b20218538