Seek to end after duration change in HTMLMediaElement https://bugs.webkit.org/show_bug.cgi?id=95986
Reviewed by Eric Carlson.

This seeks the media element to the end of the media resource if the current playback
position has become greater than the duration after a duration change.
http://dev.w3.org/html5/spec/media-elements.html#durationChange

Source/WebCore:

Test: http/tests/media/media-source/seek-to-end-after-duration-change.html

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::mediaPlayerDurationChanged): After firing the duration change event, checks to see if current time exceeds duration and if so, seeks to duration.

LayoutTests:

* http/tests/media/media-source/seek-to-end-after-duration-change-expected.txt: Added.
* http/tests/media/media-source/seek-to-end-after-duration-change.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@127949 bbb929c8-8fbe-4397-9dbb-9b2b20218538