[GTK] Remove color codes characters from run-gtk-tests output when not writing to a TTY https://bugs.webkit.org/show_bug.cgi?id=97660
Reviewed by Philippe Normand.

Remove also the '\r' characters returned by TTY output to try to fix
the double line endings of the script output in the bots.

* Scripts/run-gtk-tests:
(TestRunner.__init__): Compile a regexp for color codes if stdout
it not a TTY.
(TestRunner._run_test_command.parse_line): Remove color codes and
'\r' characters from the line returned by the pseudo-terminal if
stdout is not a TTY.
* gtk/common.py:
(parse_output_lines): Fix the last chunk check to make sure we
always send entire lines to the callback, otherwise color codes
regexp might fail if the line is not complete.

git-svn-id: svn://svn.chromium.org/blink/trunk@129627 bbb929c8-8fbe-4397-9dbb-9b2b20218538