Changed debug-safari, debug-minibrowser and debug-test-runner to use LLDB by default when using Xcode 4.5 or later.
Reviewed by Anders Carlsson.

* Scripts/webkitdirs.pm:
(determineDebugger): Changed the default debugger from GDB to LLDB when the Xcode version
is 4.5 or later. Added a --use-gdb switch for overriding this.
(printHelpAndExitForRunAndDebugWebKitAppIfNeeded): Updated to mention --use-gdb and the
how the default debugger choice depends on the Xcode version. Also changed the description
of --no-saved-state to use the same terminology Xcode’s scheme editor uses to describe this
option, and clarified to which OS X versions it applies.



git-svn-id: svn://svn.chromium.org/blink/trunk@130924 bbb929c8-8fbe-4397-9dbb-9b2b20218538