In Skia's ImageFrame, only set the isOpaque flag when the frame is complete https://bugs.webkit.org/show_bug.cgi?id=99326
Reviewed by Stephen White.

No new tests -- existing layouttests exercise this, esp.
  png-partial-load-no-alpha.html
  webp-partial-load.html
These two fail if skia is told that these are opaque, which it is w/o this CL.

At the moment, Skia has a hack to ignore the opaque-setting, so that these tests will pass as is.
This change is to first, correct webkit to only set isOpaque when the frame is complete, so that
in a later change, Skia can remove the hack, and re-enable its opaqueness optimization.

* platform/image-decoders/skia/ImageDecoderSkia.cpp:
(WebCore::ImageFrame::ImageFrame):
(WebCore::ImageFrame::operator=):
(WebCore::ImageFrame::zeroFillPixelData):
(WebCore::ImageFrame::hasAlpha):
(WebCore::ImageFrame::setHasAlpha):
(WebCore::ImageFrame::setStatus):


git-svn-id: svn://svn.chromium.org/blink/trunk@131370 bbb929c8-8fbe-4397-9dbb-9b2b20218538