[GTK] accessibility/radio-button-title-label.html is failing https://bugs.webkit.org/show_bug.cgi?id=98374
Unreviewed. Unskipped test, new port-specific baseline.

New baseline for the GTK port. If there is a proper label tag, we
expect the widget to get its name/title from that label. The aria
labelledby property is exposed in addition to the label; not instead
of the label.

Patch by Joanmarie Diggs <jdiggs@igalia.com> on 2012-10-16

* platform/gtk/TestExpectations: Unskipped failing test.
* platform/gtk/accessibility/radio-button-title-label-expected.txt: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@131510 bbb929c8-8fbe-4397-9dbb-9b2b20218538