https://bugs.webkit.org/show_bug.cgi?id=100809 ScrollingCoordinator::attachToStateTree() should take an enum indicating what type of node to create
Reviewed by Simon Fraser.

New enum ScrollingNodeType will soon have other values, such as 
FixedNode.
* page/scrolling/ScrollingCoordinator.h:
(WebCore::ScrollingCoordinator::attachToStateTree):
* page/scrolling/mac/ScrollingCoordinatorMac.h:
(ScrollingCoordinatorMac):

This patch generalizes this function so that it can be used to create 
different types of nodes and so that it is actually ready to create 
child nodes. 
* page/scrolling/mac/ScrollingCoordinatorMac.mm:
(WebCore::ScrollingCoordinatorMac::attachToStateTree):

The rootStateNode is always a ScrollingNode.
(WebCore::ScrollingCoordinatorMac::ensureRootStateNodeForFrameView):

For the time being, always create a ScrollingNode.
* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::attachToScrollingCoordinator):



git-svn-id: svn://svn.chromium.org/blink/trunk@132981 bbb929c8-8fbe-4397-9dbb-9b2b20218538