Sparse array size threshold should be increased to 100000 https://bugs.webkit.org/show_bug.cgi?id=100827
Reviewed by Oliver Hunt.

This enables the use of contiguous arrays in programs that previously
couldn't use them. And I so far can't see any examples of this being
a downside. To the extent that there is a downside, it ought to be
addressed by GC: https://bugs.webkit.org/show_bug.cgi?id=100828

* runtime/ArrayConventions.h:
(JSC):



git-svn-id: svn://svn.chromium.org/blink/trunk@133129 bbb929c8-8fbe-4397-9dbb-9b2b20218538