[Chromium] Enable the JavaScript i18n API for Android https://bugs.webkit.org/show_bug.cgi?id=101357
Reviewed by Adam Barth.

Source/WebCore:

Remove the conditional dependency on the v8-i18n API. The library is now
always available in a Chromium for Android checkout.

* WebCore.gyp/WebCore.gyp:

Source/WebKit/chromium:

Enable the JavaScript i18n API for Chromium on Android. This adds a
grand total of 18.7 KiB of data to the content shell apk.

* features.gypi:


git-svn-id: svn://svn.chromium.org/blink/trunk@133642 bbb929c8-8fbe-4397-9dbb-9b2b20218538